refactor(mobile): collapse agent-notification subsystem (16 → 8 files)
The `agent-notification-*` cluster in `apps/mobile/src/lib/` had grown to
15 source files and one component (2,212 LOC) across several iterations.
The `-core`/`-api`/`-runtime`/`-bridge-manager-core` suffixes were
iteration scars from extraction-then-rewrap rather than real boundaries.

Reshape against the intended end state, preserving:
  - the external API to `shell/detail.tsx` and `auto-connect.tsx`
  - the native (Kotlin JNI) and pure-domain (events, visibility, state
    machine) boundaries that are real
  - the test seam where module load must avoid pulling in `react-native`
    (so the route module stays split: pure factory + native singleton)

Renames:
  - `agent-notifications-native.ts` -> `agent-notification-native.ts`
    (drop misleading 's')
  - `agent-notification-runtime.ts` -> `foreground-service-runtime.ts`
    (the file owned foreground-service policy predicates, not
    notification logic; consumed only by `auto-connect.tsx`)

Test changes follow the source: bridge-runtime tests merged into the
bridge test file; route-api/runtime/native-plugin tests renamed to
match their modules.

Also fixes the preexisting `no-redeclare` lint errors on TS function
overload signatures by enabling `@typescript-eslint/no-redeclare`
(overload-aware) in place of the base rule for `.ts`/`.tsx` files.

Net: -950 LOC. typecheck clean, 381/381 integration tests pass,
0 lint errors/warnings.
