--- apps/mobile/src/lib/agent-notification-route-identity.ts
@@ -1,21 +1,0 @@
-export type AgentNotificationRouteIdentity = {
-	connectionId: string;
-	session: string;
-	windowId: string;
-	eventId: string;
-};
-
-export type AgentNotificationRouteToken = AgentNotificationRouteIdentity & {
-	tapToken: string;
-};
-
-export function createAgentNotificationRouteIdentityKey(
-	input: AgentNotificationRouteIdentity,
-) {
-	return JSON.stringify([
-		input.connectionId,
-		input.session,
-		input.windowId,
-		input.eventId,
-	]);
-}
