+++ apps/mobile/src/lib/agent-notification-route.ts
@@ -1,8 +1,29 @@
 import {
-	createAgentNotificationRouteIdentityKey,
-	type AgentNotificationRouteIdentity,
-	type AgentNotificationRouteToken,
-} from './agent-notification-route-identity';
+	createAgentNotificationPendingKey,
+	createStableNotificationId,
+} from './agent-notification-events';
+
+export type AgentNotificationRouteIdentity = {
+	connectionId: string;
+	session: string;
+	windowId: string;
+	eventId: string;
+};
+
+export type AgentNotificationRouteToken = AgentNotificationRouteIdentity & {
+	tapToken: string;
+};
+
+export function createAgentNotificationRouteIdentityKey(
+	input: AgentNotificationRouteIdentity,
+) {
+	return JSON.stringify([
+		input.connectionId,
+		input.session,
+		input.windowId,
+		input.eventId,
+	]);
+}
 
 type AgentNotificationRouteRecord = AgentNotificationRouteToken & {
 	createdAtMs: number;
@@ -24,8 +45,7 @@ export type AgentNotificationRouteTokenStoreDependencies = {
 const tokenPrefix = 'token:';
 const consumedTokenPrefix = 'consumed-token:';
 const routePrefix = 'route:';
-export const AGENT_NOTIFICATION_ROUTE_TOKEN_TTL_MS =
-	24 * 60 * 60 * 1_000;
+export const AGENT_NOTIFICATION_ROUTE_TOKEN_TTL_MS = 24 * 60 * 60 * 1_000;
 
 function tokenKey(tapToken: string) {
 	return `${tokenPrefix}${tapToken}`;
@@ -362,3 +382,57 @@ export function createAgentNotificationRouteTokenStore({
 		},
 	};
 }
+
+export type RoutedAgentNotificationAcknowledgeDependencies = {
+	deleteRouteTokens: (input: {
+		connectionId: string;
+		session: string;
+		windowId: string;
+	}) => void;
+	acknowledgeBridge: (
+		connectionId: string,
+		session: string,
+		windowId: string,
+	) => number[];
+	cancelNotification: (notificationId: number) => void;
+	warn: (message: string, error: unknown) => void;
+};
+
+export function createRoutedAgentNotificationId(
+	connectionId: string,
+	session: string,
+	windowId: string,
+) {
+	return createStableNotificationId(
+		createAgentNotificationPendingKey({ connectionId, session, windowId }),
+	);
+}
+
+export function acknowledgeRoutedAgentNotificationWithDependencies(
+	dependencies: RoutedAgentNotificationAcknowledgeDependencies,
+	input: { connectionId: string; session: string; windowId: string },
+) {
+	try {
+		dependencies.deleteRouteTokens(input);
+	} catch (error) {
+		dependencies.warn('agent notification route token cleanup failed', error);
+	}
+	const notificationIds = dependencies.acknowledgeBridge(
+		input.connectionId,
+		input.session,
+		input.windowId,
+	);
+	for (const notificationId of notificationIds) {
+		dependencies.cancelNotification(notificationId);
+	}
+	if (notificationIds.length === 0) {
+		dependencies.cancelNotification(
+			createRoutedAgentNotificationId(
+				input.connectionId,
+				input.session,
+				input.windowId,
+			),
+		);
+	}
+}
+
